pyproject.toml
[project]
name = "nse-navigators-trade-agent"
version = "0.1.0"
description = "NSE Navigators is an intelligent multi-agent trading system featuring four specialized AI agents that autonomously research, analyze, and trade Indian stocks. The first agent scouts market opportunities through news analysis and trend identification, while subsequent agents perform comprehensive financial, fundamental, and technical analysis using integrated MCP tools and servers. Each agent operates with distinct trading strategies and risk management protocols, executing trades based on their specialized decision-making frameworks. The entire process unfolds through an intuitive live dashboard where users can monitor real-time agent activities, trade executions, and portfolio performance. This fully autonomous system eliminates manual intervention while providing complete transparency and control over AI-driven investment decisions."
readme = "README.md"
authors = [
    { name = "Darshan Ramani", email = "Darshan.ramani15@gmail.com" }
    ]
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.49.0",
    "bs4>=0.0.2",
    "gradio>=5.22.0",
    "httpx>=0.28.1",
    "ipywidgets>=8.1.5",
    "lxml>=5.3.1",
    "mcp-server-fetch>=2025.1.17",
    "mcp[cli]>=1.5.0",
    "openai>=1.68.2",
    "openai-agents>=0.0.15",
    "playwright>=1.51.0",
    "plotly>=6.0.1",
    "polygon-api-client>=1.14.5",
    "psutil>=7.0.0",
    "pypdf>=5.4.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.0.1",
    "redis>=5.0.0",
    "requests>=2.32.3",
    "semantic-kernel>=1.25.0",
    "sendgrid>=6.11.0",
    "setuptools>=78.1.0",
    "speedtest-cli>=2.1.3",
    "smithery>=0.1.0",
    "pydantic",
    "pandas",
    "dhanhq",
    "ta-lib>=0.6.4",
    "apscheduler>=3.11.0",
    "scikit-learn>=1.7.0",
]
